requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
mysql-connector-python>=8.3.0
groq>=0.9.0
requests>=2.31.0
python-dotenv>=1.0.0
feedparser>=6.0.11
websocket-client>=1.7.0
